+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [unreleased]
 
+### Features
+
+- Add `TiledParallaxCamera` camera marker component to choose the camera used for map layers parallax calculations.
+
 ## v0.9.0
 
 **BREAKING CHANGES**